Move rem_scalar to utils.rs
